refactor: extract composer drafts from AgentWorkspace
